requirements.txt
Django>=2.0,<3.0
psycopg2>=2.7,<3.0
djangorestframework
djangorestframework-simplejwt
django-oauth-toolkit
django-rest-registration
django-cors-headers
django-filter
drf-yasg
django_elasticsearch_dsl
django_elasticsearch_dsl_drf
